backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
sqlalchemy
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-dotenv
pydantic
openai
openpyxl
langchain-google-genai
